src/requirements.txt
crewai[tools,agentops] 
pydantic_settings == 2.12.0
tavily-python
litellm
scrapegraph-py
